+++ src/test/java/spring/workshop/expenses/unit/Category/CategoryControllerMockTests.java
@@ -1,4 +1,4 @@
-package spring.workshop.expenses.unit;
+package spring.workshop.expenses.unit.Category;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
